% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{classification.all}
\alias{classification.all}
\title{Apply all classification methods to a precision object}
\usage{
classification.all(object, threshold_method = "cv")
}
\arguments{
\item{object}{A \code{\link{precision}} object containing harmonized training
and testing data in the slots \code{harmon.train.data},
and \code{harmon.test1.data} and \code{harmon.test2.data}, respectively.}

\item{threshold_method}{A character string specifying the thresholding method.
Options are:
- \code{"cv"}: Use cross-validation to optimize the threshold.
- \code{"none"}: Use all genes without applying a threshold.}
}
\value{
Updated precision object with all classification results added to
  the \code{classification.result} slot.
}
\description{
This function applies all classification methods to the training data
contained in a precision object.
The classification results are added to the corresponding \code{classification.result}
slot in the object for each harmonized training and testing dataset.
}
