% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_classification.R
\name{classification.lasso}
\alias{classification.lasso}
\title{LASSO Classification Function}
\usage{
classification.lasso(object, kfold = 5)
}
\arguments{
\item{object}{Input object containing harmonized data}

\item{kfold}{Number of folds for cross-validation}
}
\value{
Updated object with LASSO classification results
}
\description{
LASSO Classification Function
}
