% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_classification.R
\name{classification.ranfor}
\alias{classification.ranfor}
\title{Random Forest Classification Function}
\usage{
classification.ranfor(object, fold = 5)
}
\arguments{
\item{object}{Input object containing harmonized data}

\item{kfold}{Number of folds for cross-validation}
}
\value{
Updated object with Random Forest classification results
}
\description{
Random Forest Classification Function
}
