% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{classification.svm}
\alias{classification.svm}
\title{Support Vector Machine (SVM) Classification}
\usage{
classification.svm(object, threshold_method = "cv", kfold = 5)
}
\arguments{
\item{object}{A \link{precision} object containing harmonized data.
Must contain the slots \code{harmon.train.data} with harmonized training data
and \code{harmon.test1.data} and \code{harmon.test2.data} with harmonized test data.}

\item{threshold_method}{A character string specifying the thresholding method.
Use \code{"cv"} for cross-validation to determine the optimal k,
or \code{"none"} to use a fixed k = 1.}

\item{kfold}{An integer specifying the number of folds for cross-validation.
This parameter is only used if \code{threshold_method} is set to \code{"cv"}.}
}
\value{
The input object updated with SVM classification results added to the
 \code{classification.result} slot, including predicted classes and
 associated metrics.
}
\description{
This function applies SVM classification to the harmonized data in the
input \link{precision} object containing.
It supports two thresholding methods: cross-validation to optimize the cost parameter
or a fixed cost value of 1.
}
