% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\name{cleanup.augmented.data}
\alias{cleanup.augmented.data}
\title{Clean Up Cached Augmented Data}
\usage{
cleanup.augmented.data()
}
\value{
NULL. Used for side effects.
}
\description{
This function deletes the cached augmented data file from both the
persistent storage and temporary directory.
}
\examples{
\dontrun{
# Download and cache the augmented data
augmented.data <- load.augmented.data()
cleanup.augmented.data()
}
}
\seealso{
\code{\link{load.augmented.data}} for loading the data.
}
