% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{cluster.all}
\alias{cluster.all}
\title{Apply all clustering methods to a precision object}
\usage{
\method{cluster}{all}(object, k = NULL)
}
\arguments{
\item{object}{A \code{\link{precision}} object containing harmonized training data
in the slot \code{harmon.train.data}.}

\item{k}{\emph{Integer.} Number of clusters. If NULL or not given,
uses the number of unique labels in training data.}
}
\value{
Updated precision object with all clustering results added to
the \code{cluster.result} slot.
}
\description{
This function applies all clustering methods to the training data
contained in a precision object.
The clustering results are added to the corresponding \code{cluster.result}
slot in the object for each training dataset.
}
\details{
The following clustering methods are applied to the data:
  \itemize{
    \item Hierarchical Clustering using multiple distance metrics
      (Euclidean, Pearson, and Spearman)
      (\code{\link{cluster.hc}})
    \item Gaussian Mixture Models (MNM) (\code{\link{cluster.mnm}})
    \item K-Means Clustering (\code{\link{cluster.kmeans}})
    \item Self-Organizing Maps (\code{\link{cluster.som}})
    \item Partitioning Around Medoids (Euclidean, Pearson, and Spearman)
      (\code{\link{cluster.pam}})
  }
}
