% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_clustering.R
\name{cluster.kmeans}
\alias{cluster.kmeans}
\title{Perform K-means clustering on harmonized data}
\usage{
cluster.kmeans(object, k = NULL)
}
\arguments{
\item{object}{An object containing harmonized training data}

\item{k}{Integer. Number of clusters. If NULL, uses the number of unique labels in training data}
}
\value{
Updated object with k-means clustering results
}
\description{
Perform K-means clustering on harmonized data
}
