% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{cluster.mnm}
\alias{cluster.mnm}
\title{Perform Gaussian Mixture Model clustering on harmonized data}
\usage{
cluster.mnm(object, k = NULL)
}
\arguments{
\item{object}{An object containing harmonized training data}

\item{k}{Integer. Number of clusters. If NULL, uses the number of unique labels in training data}
}
\value{
Updated object with Gaussian Mixture Model clustering results
}
\description{
Perform Gaussian Mixture Model clustering on harmonized data
}
