% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_clustering.R
\name{cluster.pam.pearson}
\alias{cluster.pam.pearson}
\title{Perform PAM clustering with Pearson correlation distance on harmonized data}
\usage{
cluster.pam.pearson(object, k = NULL)
}
\arguments{
\item{object}{An object containing harmonized training data}

\item{k}{Integer. Number of clusters. If NULL, uses the number of unique labels in training data}
}
\value{
Updated object with PAM clustering results using Pearson correlation distance
}
\description{
Perform PAM clustering with Pearson correlation distance on harmonized data
}
