% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_clustering.R
\name{cluster.pam.spearman}
\alias{cluster.pam.spearman}
\title{Perform PAM clustering with Spearman correlation distance on harmonized data}
\usage{
cluster.pam.spearman(object, k = NULL)
}
\arguments{
\item{object}{An object containing harmonized training data}

\item{k}{Integer. Number of clusters. If NULL, uses the number of unique labels in training data}
}
\value{
Updated object with PAM clustering results using Spearman correlation distance
}
\description{
Perform PAM clustering with Spearman correlation distance on harmonized data
}
