% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{cluster.som}
\alias{cluster.som}
\title{Perform Self-Organizing Map (SOM) clustering on harmonized data}
\usage{
cluster.som(object, k = NULL)
}
\arguments{
\item{object}{An object containing harmonized training data}

\item{k}{Integer. Number of clusters. If NULL, uses the number of unique labels in training data}
}
\value{
Updated object with SOM clustering results
}
\description{
Perform Self-Organizing Map (SOM) clustering on harmonized data
}
