% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonization.R
\name{create.precision.classification}
\alias{create.precision.classification}
\title{Create precision object for classification}
\usage{
create.precision.classification(
  traindata,
  testdata1,
  testdata2,
  trainlabel,
  testlabel1,
  testlabel2
)
}
\arguments{
\item{traindata}{Training data matrix, (\code{p} by \code{n}), where
\code{p} is the number of features and \code{n} is the number of samples.}

\item{testdata1}{First test data matrix, (\code{p} by \code{n}), where
\code{p} is the number of features and \code{n} is the number of samples.}

\item{testdata2}{Second test data matrix, (\code{p} by \code{n}), where
\code{p} is the number of features and \code{n} is the number of samples.}

\item{trainlabel}{Training data labels, a vector of length \code{n} (number of samples).}

\item{testlabel1}{First test data labels, a vector of length \code{n} (number of samples).}

\item{testlabel2}{Second test data labels, a vector of length \code{n} (number of samples).}
}
\value{
A precision object for classification
}
\description{
Use this function to create a \code{\link{precision}} object for classification tasks.
This object will contain raw training and two sets of test data,
along with their labels.
}
