% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonization.R
\name{create.precision.cluster}
\alias{create.precision.cluster}
\title{Create precision object for clustering}
\usage{
create.precision.cluster(data, label)
}
\arguments{
\item{data}{Data matrix}

\item{label}{Labels}
}
\value{
A precision object for clustering
}
\description{
Use this function to create a \code{\link{precision}} object for clustering tasks.
This object will contain raw training data and its labels.
}
