% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_effects.R
\name{handling.effects}
\alias{handling.effects}
\title{Add handling effects to benchmark data}
\usage{
handling.effects(clean.input, benchmark, test, group, d)
}
\arguments{
\item{clean.input}{Numeric matrix of clean input data}

\item{benchmark}{Numeric matrix of benchmark data}

\item{test}{Numeric matrix of test data}

\item{group}{Factor or vector defining sample groups}

\item{d}{Numeric effect size factor}
}
\value{
List containing modified data and group information
}
\description{
Add handling effects to benchmark data
}
