% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonization.R
\name{harmon.all}
\alias{harmon.all}
\title{Apply all harmonization methods to a precision object}
\usage{
harmon.all(object, add.raw = TRUE)
}
\arguments{
\item{object}{A \link{precision} object containing raw data.
Harmonization is applied to the slots
\code{raw.train.data}, \code{raw.test1.data}, and \code{raw.test2.data}
(if applicable), each containing a list with \code{data} and \code{label} elements.}

\item{add.raw}{Logical. If TRUE, add the raw data to the harmonized data}
}
\value{
A precision object with harmonization results added to the
  slots \code{harmon.train.data}, \code{harmon.test1.data}, and \code{harmon.test2.data}
  (if applicable).
}
\description{
This function applies all harmonization methods to the raw data
contained in a precision object.
The harmonization results are added to the corresponding slots in the
object for each dataset ("train", "test1", "test2") if they exist.
}
\details{
The following harmonization methods are applied to the data:
  \itemize{
    \item No Harmonization (if \code{add.raw = TRUE})
    \item Total Count (TC) (\code{\link{harmon.TC}})
    \item Upper Quartile (UQ) (\code{\link{harmon.UQ}})
    \item Median (median) (\code{\link{harmon.med}})
    \item Trimmed Median of Means (TMM) (\code{\link{harmon.TMM}})
    \item DESeq (\code{\link{harmon.DESeq}})
    \item PoissonSeq (\code{\link{harmon.PoissonSeq}})
    \item Quantile Normalization (QN) (\code{\link{harmon.QN}})
    \item Remove Unwanted Variation (RUVg, RUVr, and RUVs)
      (\code{\link{harmon.RUVg}}), (\code{\link{harmon.RUVr}}),
   (\code{\link{harmon.RUVs}})
  }
}
