% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonization.R
\name{harmon.DESeq}
\alias{harmon.DESeq}
\title{Apply DESeq harmonization}
\usage{
harmon.DESeq(object)
}
\arguments{
\item{object}{A precision object containing raw data}
}
\value{
A precision object with DESeq harmonization results added
}
\description{
Apply DESeq harmonization
}
