% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonization.R
\name{harmon.med}
\alias{harmon.med}
\title{Apply median harmonization}
\usage{
harmon.med(object)
}
\arguments{
\item{object}{A precision object containing raw data}
}
\value{
A precision object with median harmonization results added
}
\description{
Apply median harmonization
}
