% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonization.R
\name{harmon.PoissonSeq}
\alias{harmon.PoissonSeq}
\title{Apply PoissonSeq harmonization}
\usage{
harmon.PoissonSeq(object)
}
\arguments{
\item{object}{A precision object containing raw data}
}
\value{
A precision object with PoissonSeq harmonization results added
}
\description{
Apply PoissonSeq harmonization
}
