% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonization.R
\name{harmon.QN.frozen}
\alias{harmon.QN.frozen}
\title{Apply Standard Quantile normalization with frozen parameters}
\usage{
harmon.QN.frozen(object)
}
\arguments{
\item{object}{A \link{precision} object containing raw data.
Harmonization is applied to the slots
\code{raw.train.data}, \code{raw.test1.data}, and \code{raw.test2.data}
(if applicable), each containing a list with \code{data} and \code{label} elements.}
}
\value{
A precision object with QN harmonization results added to the
  slots \code{harmon.train.data}, \code{harmon.test1.data}, and \code{harmon.test2.data}
  (if applicable).
}
\description{
Quantile normalization for miRNA-Seq data using the preprocessCore package.
For more information, visit documentation on the
\href{https://bioconductor.org/packages/release/bioc/html/preprocessCore.html}{preprocessCore Bioconductor page}.
}
\details{
This function applies quantile normalization with frozen parameters to the
raw data contained in a precision object. The harmonization results are added
to the corresponding slots in the object for each dataset
("train", "test1", "test2") if they exist.
}
