% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_basic_harmonization_functions.R
\name{harmon.QN}
\alias{harmon.QN}
\title{Apply standard quantile normalization}
\usage{
harmon.QN(object)
}
\arguments{
\item{object}{A precision object containing raw data}
}
\value{
A precision object with quantile normalization (QN) results added
}
\description{
Apply standard quantile normalization
}
