% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonization.R
\name{harmon.SVA}
\alias{harmon.SVA}
\title{Apply SVA harmonization}
\usage{
harmon.SVA(object)
}
\arguments{
\item{object}{A precision object containing raw data}
}
\value{
A precision object with SVA harmonization results added
}
\description{
Apply SVA harmonization
}
