% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_basic_harmonization_functions.R
\name{harmon.TMM.frozen}
\alias{harmon.TMM.frozen}
\title{Apply TMM normalization with frozen parameters}
\usage{
harmon.TMM.frozen(object)
}
\arguments{
\item{object}{A precision object containing raw data}
}
\value{
A precision object with TMM frozen harmonization results added
}
\description{
Apply TMM normalization with frozen parameters
}
