% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_harmonization.R
\docType{class}
\name{precision-class}
\alias{precision-class}
\alias{precision}
\title{Precision Class}
\description{
Precision Class
}
\section{Slots}{

\describe{
\item{\code{raw.train.data}}{Training data and labels}

\item{\code{raw.test1.data}}{First test dataset and labels}

\item{\code{raw.test2.data}}{Second test dataset and labels}

\item{\code{harmon.train.data}}{List of harmonized training data}

\item{\code{harmon.test1.data}}{List of harmonized test1 data}

\item{\code{harmon.test2.data}}{List of harmonized test2 data}

\item{\code{classification.result}}{Classification results}

\item{\code{cluster.result}}{Clustering results}
}}

