% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{tabulate.ext.err.func}
\alias{tabulate.ext.err.func}
\title{Calculate external error rate}
\usage{
tabulate.ext.err.func(pred.obj, obs.grp)
}
\arguments{
\item{pred.obj}{Predicted values}

\item{obs.grp}{Observed groups}
}
\value{
Error rate
}
\description{
Calculate external error rate
}
