% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_effects.R
\name{biological.effects}
\alias{biological.effects}
\title{Add biological effects to benchmark data}
\usage{
biological.effects(benchmark, group, c)
}
\arguments{
\item{benchmark}{Numeric matrix of expression values}

\item{group}{Factor or vector defining sample groups}

\item{c}{Numeric amplification factor}
}
\value{
List containing modified data and group information
}
\description{
Add biological effects to benchmark data
}
