% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_classification.R
\name{classification.knn}
\alias{classification.knn}
\title{KNN Classification Function}
\usage{
classification.knn(object, threshold_method = "cv", kfold = 5, folds = NULL)
}
\arguments{
\item{object}{Input object containing harmonized data}

\item{threshold_method}{Character. Either "cv" for cross-validation optimized k or "none" for fixed k=1}

\item{kfold}{Number of folds for cross-validation (only used if threshold_method = "cv")}

\item{folds}{Pre-specified folds (optional)}
}
\value{
Updated object with KNN classification results
}
\description{
KNN Classification Function
}
