% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_classification.R
\name{classification.pam}
\alias{classification.pam}
\title{Main PAM classification function}
\usage{
classification.pam(object, vt.k = NULL, n.k = 30, kfold = 5, folds = NULL)
}
\arguments{
\item{object}{Input object containing harmonized data}

\item{vt.k}{Threshold values}

\item{n.k}{Number of threshold values}

\item{kfold}{Number of folds}
}
\value{
Updated object with classification results
}
\description{
Main PAM classification function
}
