% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_classification.R
\name{classification.pam}
\alias{classification.pam}
\title{Main PAM classification function}
\usage{
classification.pam(object, threshold_method = "cv", vt.k = NULL, kfold = 5)
}
\arguments{
\item{object}{Input object containing harmonized data}

\item{threshold_method}{Character. Either "cv" for cross-validation optimized threshold or "none" for using all genes}

\item{vt.k}{Threshold values (only used if threshold_method = "cv")}

\item{kfold}{Number of folds}
}
\value{
Updated object with classification results
}
\description{
Main PAM classification function
}
