% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{classification.ranfor}
\alias{classification.ranfor}
\title{Random Forest Classification Function}
\usage{
classification.ranfor(object, threshold_method = "cv", kfold = 5)
}
\arguments{
\item{object}{Input object containing harmonized data}

\item{threshold_method}{Character. Either "cv" for cross-validation tuned parameters or "none" for default parameters}

\item{kfold}{Number of folds for cross-validation (only used if threshold_method = "cv")}
}
\value{
Updated object with Random Forest classification results
}
\description{
Random Forest Classification Function
}
