% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_classification.R
\name{classification.svm}
\alias{classification.svm}
\title{SVM Classification Function}
\usage{
classification.svm(object, threshold_method = "cv", kfold = 5)
}
\arguments{
\item{object}{Input object containing harmonized data}

\item{threshold_method}{Character. Either "cv" for cross-validation tuned parameters or "none" for default parameters}

\item{kfold}{Number of folds for cross-validation (only used if threshold_method = "cv")}
}
\value{
Updated object with SVM classification results
}
\description{
SVM Classification Function
}
