% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_classification.R
\name{classification.svm}
\alias{classification.svm}
\title{SVM Classification Function}
\usage{
classification.svm(object, kfold = 5)
}
\arguments{
\item{object}{Input object containing harmonized data}

\item{kfold}{Number of folds for cross-validation}
}
\value{
Updated object with SVM classification results
}
\description{
SVM Classification Function
}
