% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_clustering.R
\name{cluster.hc}
\alias{cluster.hc}
\title{Perform hierarchical clustering on harmonized data}
\usage{
cluster.hc(object, k = NULL)
}
\arguments{
\item{object}{An object containing harmonized training data}

\item{k}{Integer. Number of clusters. If NULL, uses the number of unique labels in training data}
}
\value{
Updated object with hierarchical clustering results
}
\description{
Perform hierarchical clustering on harmonized data
}
