% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{cluster.kmeans}
\alias{cluster.kmeans}
\title{K-means clustering for harmonized data}
\usage{
\method{cluster}{kmeans}(object, k = NULL)
}
\arguments{
\item{object}{A \code{\link{precision}} object containing harmonized training data
in the slot \code{harmon.train.data}.}

\item{k}{\emph{Integer.} Number of clusters. If NULL or not given,
uses the number of unique labels in training data.}
}
\value{
Updated precision object with k-means clustering results added to
the \code{cluster.result} slot.
}
\description{
This function performs K-means clustering on the harmonized training data
and returns the clustering results.
}
