% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{cluster.mnm}
\alias{cluster.mnm}
\title{Gaussian Mixture Model clustering for harmonized data}
\usage{
cluster.mnm(object, k = NULL)
}
\arguments{
\item{object}{A \code{\link{precision}} object containing harmonized training data
in the slot \code{harmon.train.data}.}

\item{k}{\emph{Integer.} Number of clusters. If NULL or not given,
uses the number of unique labels in training data.}
}
\value{
Updated precision object with Gaussian Mixture Model clustering
results added to the \code{cluster.result} slot.
}
\description{
This function performs Gaussian Mixture Model (GMM) clustering on the harmonized
training data using the `mclust` package.
}
