% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{cluster.pam}
\alias{cluster.pam}
\title{Perform PAM clustering with specified distance measure on harmonized data}
\usage{
cluster.pam(object, k = NULL, distance = "euclidean")
}
\arguments{
\item{object}{An object containing harmonized training data}

\item{k}{Integer. Number of clusters. If NULL, uses the number of unique labels in training data}

\item{distance}{Character. Distance measure to use: "euclidean", "pearson", or "spearman"}
}
\value{
Updated object with PAM clustering results using specified distance measure
}
\description{
Perform PAM clustering with specified distance measure on harmonized data
}
