% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_basic_harmonization_functions.R
\name{create.precision.classification}
\alias{create.precision.classification}
\title{Create precision object for classification}
\usage{
create.precision.classification(
  traindata,
  testdata1,
  testdata2,
  trainlabel,
  testlabel1,
  testlabel2
)
}
\arguments{
\item{traindata}{Training data matrix}

\item{testdata1}{First test data matrix}

\item{testdata2}{Second test data matrix}

\item{trainlabel}{Training data labels}

\item{testlabel1}{First test data labels}

\item{testlabel2}{Second test data labels}
}
\value{
A precision object
}
\description{
Create precision object for classification
}
