% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_basic_harmonization_functions.R
\name{create.precision.cluster}
\alias{create.precision.cluster}
\title{Create precision object for clustering}
\usage{
create.precision.cluster(data, label)
}
\arguments{
\item{data}{Data matrix}

\item{label}{Labels}
}
\value{
A precision object
}
\description{
Create precision object for clustering
}
