% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_effects.R
\name{handling.effects}
\alias{handling.effects}
\title{Add handling effects to benchmark data}
\usage{
handling.effects(clean.input, benchmark, test, group, d)
}
\arguments{
\item{clean.input}{Numeric matrix of clean input data, (\code{p} by \code{n}), where
\code{p} is the number of features and \code{n} is the number of samples.}

\item{benchmark}{Numeric matrix of benchmark data, (\code{p} by \code{n}), where
\code{p} is the number of features and \code{n} is the number of samples.}

\item{test}{Numeric matrix of test data, (\code{p} by \code{n}), where
\code{p} is the number of features and \code{n} is the number of samples.}

\item{group}{Factor or vector of character strings defining sample groups
for each sample. Must have exactly two levels.}

\item{d}{Numeric effect strength factor}
}
\value{
List containing modified data, group information, and effect size
}
\description{
This function simulates handling effects by modifying the
clean input data based on the differences between the benchmark and test data.
}
