% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_basic_harmonization_functions.R
\name{harmon.ComBat.Seq}
\alias{harmon.ComBat.Seq}
\title{Apply ComBat-Seq normalization}
\usage{
harmon.ComBat.Seq(object, batches)
}
\arguments{
\item{object}{A precision object containing raw data}

\item{batches}{A vector of batch labels for the samples}
}
\value{
A precision object with ComBat-Seq normalization results added
}
\description{
Apply ComBat-Seq normalization
}
