% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_basic_harmonization_functions.R
\name{harmon.QN.frozen}
\alias{harmon.QN.frozen}
\title{Apply quantile normalization with frozen parameters}
\usage{
harmon.QN.frozen(object)
}
\arguments{
\item{object}{A precision object containing raw data}
}
\value{
A precision object with frozen quantile normalization (QN) results added
}
\description{
Apply quantile normalization with frozen parameters
}
