% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_harmonization.R
\name{harmon.RUVg}
\alias{harmon.RUVg}
\title{Apply RUVg harmonization}
\usage{
harmon.RUVg(object)
}
\arguments{
\item{object}{A precision object containing raw data}
}
\value{
A precision object with RUVg harmonization results added
}
\description{
Apply RUVg harmonization
}
