% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_basic_harmonization_functions.R
\name{harmon.RUVr}
\alias{harmon.RUVr}
\title{Apply RUVr harmonization}
\usage{
harmon.RUVr(object)
}
\arguments{
\item{object}{A precision object containing raw data}
}
\value{
A precision object with RUVr harmonization results added
}
\description{
Apply RUVr harmonization
}
