% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_harmonization.R
\name{harmon.RUVs}
\alias{harmon.RUVs}
\title{Apply RUVs harmonization}
\usage{
harmon.RUVs(object)
}
\arguments{
\item{object}{A precision object containing raw data}
}
\value{
A precision object with RUVs harmonization results added
}
\description{
Apply RUVs harmonization
}
