% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonization.R
\name{harmon.TC}
\alias{harmon.TC}
\title{Apply total count harmonization}
\usage{
harmon.TC(object)
}
\arguments{
\item{object}{A precision object containing raw data}
}
\value{
A precision object with total count (TC) harmonization results added
}
\description{
Apply total count harmonization
}
