% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonization.R
\name{harmon.TMM}
\alias{harmon.TMM}
\title{Apply standard TMM normalization}
\usage{
harmon.TMM(object)
}
\arguments{
\item{object}{A precision object containing raw data}
}
\value{
A precision object with TMM harmonization results added
}
\description{
Apply standard TMM normalization
}
