% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonization.R
\name{harmon.UQ}
\alias{harmon.UQ}
\title{Apply upper quartile harmonization}
\usage{
harmon.UQ(object)
}
\arguments{
\item{object}{A precision object containing raw data}
}
\value{
A precision object with upper quartile (UQ) harmonization results added
}
\description{
Apply upper quartile harmonization
}
